pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rivet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "httpx>=0.28.1",
    "langgraph>=1.0.4",
    "langsmith>=0.4.49",
    "litellm>=1.80.7",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "rich>=14.2.0",
    "typer>=0.20.0",
]

[project.scripts]
rivet = "rivet.main:app"

[dependency-groups]
dev = [
    "pre-commit>=4.5.0",
    "ruff>=0.14.7",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]
